인티저에 실수를 대입해서 에러????
실수를 정수로 변환함...

+++ selfdrive/car/hyundai/hyundaican.py
@@ -103,7 +103,7 @@ def create_acc_commands(packer, enabled, accel, upper_jerk, idx, lead_visible, s
   cruise_enabled = enabled and CS.out.cruiseState.enabled
   #objGap = 0 if vision_dist == 0 else 2 if vision_dist < 25 else 3 if vision_dist < 40 else 4 if vision_dist < 70 else 5 
   print('vision_dist = {}'.format(vision_dist))
-  print('cluster_speed = {}'.format(CS.vEgo*3.6))
+  print('cluster_speed = {}'.format(int(CS.vEgo*3.6)))
   
   scc11_values = {
     "MainMode_ACC": CS.out.cruiseState.available, # SCC 모드 여부...
